Lift platform restriction for pio users
On platform.io (pio) the library is currently restricted to the ESP8266
platform only. In Arduino IDE the library is unrestricted.

I do no see any reason why this limitation exists for pio users,
especially given that fact that the project's README states that it is
tested on different Arduinos, too. This also synchronizes pio settings
with settings for the Arduino IDE.