Replace inheritance for composition.
Add an AdaFruit_NeoPixel strand as a class member.

The intention is to be able to eventually pass in an AdaFruit strand to a WS2812FX.
And then be able to pass in multiple AdaFruit strands to a WS2812FX.
This will allow for a single animation to span multiple strands.