esp8266_webinterface: stop auto cycling after manual mode selection
When a user selects a mode manually, it gives a strange user experience when 5 seconds later the mode changes. After this change, auto cycling will stop when selecting a mode manually in the web interface. Of course, a user can start the auto cycling again using the auto cycle button.